requirements_original.txt
beautifulsoup4==4.14.3
datasets==4.1.1
deep_translator==1.11.4
evaluate==0.4.6
fastapi==0.128.0
google_api_python_client==2.184.0
google_auth_oauthlib==1.2.2
matplotlib==3.10.8
nltk==3.8.1
numpy==2.4.0
pandas==2.3.3
protobuf==6.33.2
pydantic==2.12.5
python-dotenv==1.2.1
scikit_learn==1.8.0
seaborn==0.13.2
torch==2.9.0
tqdm==4.67.1
transformers==4.57.3
uvicorn==0.40.0
